requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
accelerate==0.24.0

# Computer Vision
ultralytics==8.0.200
opencv-python==4.8.1.78
Pillow==10.1.0
easyocr==1.7.0

# Utils
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
numpy==1.24.3
aiofiles==23.2.0
